Median Calculator
Hey there! I created a median calculator using python 3. However, this calculator has many limitations. For example it can only work on an odd count of inputted numbers and requires a comma between each number. This project is quite basic and isn't meant to be used for practical purposes. It is just an example of creating and applying successful algorithms. I am on the road to learn python and am ready to learn. Anyone, who wants can reach out as I would love to collaborate and learn from them.